[238] Add ability to exclude Outdoor or Indoor workouts respectively
