--- src/Common/Dto/AverageSummary.cs
@@ -1,10 +1,0 @@
-﻿namespace Common.Dto
-{
-	public class AverageSummary
-	{
-		public string Display_Name { get; set; }
-		public string Display_Unites { get; set; }
-		public double? Value { get; set; }
-		public string Slug { get; set; } // enum
-	}
-}
