+++ src/Common/Dto/Peloton/PowerZones.cs
