+++ src/Common/Dto/Peloton/Summary.cs
@@ -1,0 +1,10 @@
+﻿namespace Common.Dto
+{
+	public class Summary
+	{
+		public string Display_Name { get; set; }
+		public string Display_Unit { get; set; }
+		public double? Value { get; set; }
+		public string Slug { get; set; } // enum
+	}
+}
