+++ src/Common/Dto/Peloton/Zone.cs
@@ -1,0 +1,12 @@
+﻿namespace Common.Dto
+{
+	public class Zone
+	{
+		public string Display_Name { get; set; }
+		public string Slug { get; set; }
+		public string Range { get; set; }
+		public int Duration { get; set; }
+		public double Max_Value { get; set; }
+		public double Min_Value { get; set; }
+	}
+}
