+++ src/Common/Dto/WorkoutType.cs
@@ -1,0 +1,19 @@
+﻿namespace Common.Dto
+{
+	public enum WorkoutType
+	{
+		None = 0,
+		Cycling,
+		BikeBootcamp,
+		TreadmillRunning,
+		OutdoorRunning,
+		TreadmillWalking,
+		OutdoorWalking,
+		Cardio,
+		Circuit,
+		Strength,
+		Stretching,
+		Yoga,
+		Meditation
+	}
+}
