--- src/Common/Dto/Zone.cs
@@ -1,12 +1,0 @@
-﻿namespace Common.Dto
-{
-	public class Zone
-	{
-		public string Display_Name { get; set; }
-		public string Slug { get; set; }
-		public string Range { get; set; }
-		public int Duration { get; set; }
-		public double Max_Value { get; set; }
-		public double Min_Value { get; set; }
-	}
-}
